domain_info/requirements.txt
python-whois
ip2geotools
termcolor